fixup! SOX-54 Cross branch metadata behavior tested
